fix: storage topology
The topology should reflect only the nodes that can actually access the shared storage.
Otherwise, Kubernetes may schedule a Pod onto a node that cannot attach the volume.

Unfortunately, new nodes will not be included in this list, so the Pod cannot be scheduled onto a newly added Proxmox node that uses the shared storage.

Signed-off-by: Serge Logvinov <serge.logvinov@sinextra.dev>